+++ src/Domain/Entities/Transaction.cs
@@ -4,10 +4,10 @@ namespace Domain.Entities;
 [Table("Transactions")]
 public class Transaction
 {
-    public int TransactionId { get; set; }
-    public int SourceAccountId { get; set; }
+    public long TransactionId { get; set; }
+    public long SourceAccountId { get; set; }
     public Account SourceAccount { get; set; }
-    public int DestinationAccountId { get; set; }
+    public long DestinationAccountId { get; set; }
     public Account DestinationAccount { get; set; }
     public decimal Amount { get; set; }
     public DateTime Date { get; set; }
